[examples][exp] Add SVRG and LDA config file and re-organize the code
