+++ examples/lda/pslda.ini
@@ -1,0 +1,49 @@
+worker_port=29999
+master_port=50239
+master_host=proj10
+cluster_manager_host=proj10
+cluster_manager_port=51222
+hdfs_namenode=proj10
+hdfs_namenode_port=9000
+comm_port=3245
+log_dir=/tmp/yidi/log
+
+serve=0
+
+task_scheduler_type=sequential
+scheduler_trigger_time_out=
+scheduler_trigger_num_threads=
+
+consistency=SSP
+staleness=1
+# for nytimes  
+# input=hdfs:///yidi/pre_processed_nytimes1
+# max_voc_id=102660
+# for pub_med a relatively large dataset  3.5 GB
+input=hdfs:///yidi/pre_processed_pubmed
+max_voc_id=141044
+
+num_iterations=300
+num_topics=1000
+alpha=0.1
+beta=0.1
+
+num_load_workers=100
+num_train_workers=100
+
+# the exp_result time and llh statistics write path
+result_write_path=/home/yidi/pslda_log
+
+# whether or not to compute_llh
+compute_llh=1
+num_process=1 
+num_batches=1
+max_vocs_each_pull=1000000
+
+[worker]
+#info=master:10
+info=proj5:20 
+info=proj6:20
+info=proj7:20
+info=proj8:20
+info=proj9:20
