+++ exp/CMakeLists.txt
@@ -1,5 +1,9 @@
 include_directories(${PROJECT_SOURCE_DIR} ${PROJECT_SOURCE_DIR}/husky ${HUSKY_EXTERNAL_INCLUDE})
 
+add_subdirectory(svrg)
+
+### Applications
+
 # ReadLocalGlobal
 add_executable(ReadLocalGlobal read_local_global.cpp)
 target_link_libraries(ReadLocalGlobal husky)
@@ -62,10 +66,3 @@ target_link_libraries(AutoElasticPS husky)
 target_link_libraries(AutoElasticPS husky-ml)
 target_link_libraries(AutoElasticPS ${HUSKY_EXTERNAL_LIB})
 husky_default_properties(AutoElasticPS)
-
-# re-organize LR SVRG
-add_executable(LR lrsvrg.cpp)
-target_link_libraries(LR husky)
-target_link_libraries(LR husky-ml)
-target_link_libraries(LR ${HUSKY_EXTERNAL_LIB})
-husky_default_properties(LR)
