+++ exp/svrg/lrsvrg.ini
@@ -1,0 +1,87 @@
+# system configuration
+master_host=proj10
+master_port=43150
+worker_port=44320
+comm_port=33193
+cluster_manager_host=proj10
+cluster_manager_port=45220
+hdfs_namenode=proj10
+hdfs_namenode_port=9000
+
+serve=0
+
+task_scheduler_type=autoparallelism
+# task_scheduler_type=sequential
+scheduler_trigger_time_out=1
+scheduler_trigger_num_threads=1
+
+
+# 130 MB  binary
+# input=hdfs:///ml-bin/avazu-app-part-bin
+# alpha=0.5
+# num_iters=10000
+# num_features=1000000
+# train_epoch=10
+
+# 2.22 MB
+# input=hdfs:///datasets/classification/a9
+# alpha=0.005
+# num_features=123
+# data_size=32561
+
+# 2.66 GB 
+# input=hdfs:///datasets/classification/avazu-app
+# alpha=0.5
+# num_iters=1000
+# num_features=1000000
+# train_epoch=10
+
+# 130 MB 
+# input=hdfs:///datasets/classification/avazu-app-part
+# alpha=0.5
+# num_iters=10000
+# num_features=1000000
+# train_epoch=6
+
+# 4.69 GB 
+# jinput=hdfs:///datasets/classification/avazu-site
+# jalpha=0.5
+# jnum_iters=10000
+# jnum_features=1000000
+# jtrain_epoch=10
+
+# 7.39 GB 
+# input=hdfs:///datasets/classification/higgs
+# alpha=0.05
+# num_iters=1000
+# num_features=28
+# data_size=11000000
+
+# 20.93 GB
+input=hdfs:///datasets/classification/kdd12
+alpha=0.05
+num_features=54686452
+data_size=149639105
+
+# 24 GB
+#input=hdfs:///datasets/classification/webspam3
+#alpha=1.0
+#num_features=16609143
+#data_size=350000
+
+num_load_workers=100
+num_train_workers=100
+
+fgd_threads_per_worker=15
+sgd_threads_per_worker=6
+
+sgd_overall_batch_size=1496
+outer_loop=3
+inner_loop=50
+
+[worker]
+info=proj5:20
+info=proj6:20
+info=proj7:20
+info=proj8:20
+info=proj9:20
